+++ backend/src/main/java/com/staccato/image/service/AwsS3Service.java
@@ -1,4 +1,4 @@
-package com.staccato.image.infrastructure;
+package com.staccato.image.service;
 
 import java.util.ArrayList;
 import java.util.List;
